+++ quick-protobuf/benches/perftest.rs
@@ -12,6 +12,7 @@ use perftest_data::*;
 
 use quick_protobuf::message::MessageWrite;
 use quick_protobuf::{BytesReader, Reader, Writer};
+use quick_protobuf::MessageRead;
 use test::{black_box, Bencher};
 
 #[bench]
