+++ run_test.sh
@@ -4,5 +4,8 @@ cargo run -p pb-rs quick-protobuf/examples/pb_rs/data_types_import.proto
 cargo run -p pb-rs quick-protobuf/examples/pb_rs/data_types.proto
 cargo run -p quick-protobuf --example pb_rs_example
 
+cargo run -p pb-rs quick-protobuf/benches/perftest_data/perftest_data.proto
+cargo run -p pb-rs quick-protobuf/benches/rust-protobuf/perftest_data_quick.proto
+
 cd quick-protobuf/tests/rust_protobuf
 ./generate.sh
